fix redirect crashes
